# swegger/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Test
  contact:
    email: raneendwikat16@gmail.com
  license:
    name: MIT
    url: http://www.opensource.org/license/MIT
  version: 1.0.11
...
servers:
  - url: http://localhost:3000/
paths:
...
  /api/v1/task/update/{_id}:
    put:
      tags:
        - Tasks
      summary: Update task
      description: This can only be done by the logged in user.
      operationId: updateTask
      parameters:
        - name: _id
          in: path
          description: id that need to be deleted
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent Task 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
          application/xml:
            schema:
              $ref: '#/components/schemas/Task'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        default:
          description: successful operation
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
          example: 63543c168d39fb31ddbc3a60
        name:
          type: string
          example: Mike
        salary:
          type: integer
          description: User salary
          format: int32
          example: 1500
      xml:
        name: user
    Task:
      required:
        - title
        - owner
      type: object
      properties:
        _id:
          type: string
          readOnly: true
          example: 63543c168d39fb31ddbc3a60
        title:
          type: string
          example: 1st task
        description:
          type: string
          example: 1st task description
        owner:
          $ref: '#/components/schemas/User'
        status:
          type: string
          description: task status 
          enum:
            - completed
            - pending
            - cancelled
      xml:
        name: task